backend/requirements.txt
flask
flask-cors
bcrypt
pymongo
python-dotenv
PyJWT